// provus-mobile-platform/enums/EstadoSubmissaoEnum.ts
export enum EstadoSubmissaoEnum {
  INICIADA = "Iniciada",
  ENVIADA = "Enviada",
  AVALIADA = "Avaliada",
  ENCERRADA = "Encerrada",
  ABANDONADA = "Abandonada",
  EM_ANDAMENTO = "Em Andamento",
  REABERTA = "Reaberta",
  PAUSADA = "Pausada",
  CANCELADA = "Cancelada",
  CODIGO_CONFIRMADO = "Código confirmado",
}

export default EstadoSubmissaoEnum;